+++ retail/modules/healthbarColor.lua
@@ -29,9 +29,9 @@ function BBF.SetVertexColor(frame, r, g, b, a)
 end
 
 local function getUnitReaction(unit)
-    if UnitIsFriend("player", unit) then
+    if UnitIsFriend(unit, "player") then
         return "FRIENDLY"
-    elseif UnitIsEnemy("player", unit) then
+    elseif UnitIsEnemy(unit, "player") then
         return "HOSTILE"
     else
         return "NEUTRAL"